starter-projects/django-api-postgres/pyproject.toml
[tool.poetry]
name = "django-api-postgres"
version = "0.1.0"
description = "Stacktape example Django API"
authors = ["Stacktape team <info@stacktape.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
dj-database-url = "^0.5.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
